+++ private/lib/Extension/Panel/Manager.php
@@ -499,7 +499,7 @@ final class Manager
      */
     public function stockExtensionDirectories(): array
     {
-        return ['contact', 'cron', 'database', 'phpinfo', 'repo', 'signups'];
+        return ['backup', 'contact', 'cron', 'database', 'phpinfo', 'repo', 'signups'];
     }
 
     /**